make device id variable length and use wifi mac for esp32
