+++ app/brewblox-esp/main/network/CboxConnection.cpp
@@ -13,38 +13,6 @@ CboxConnection::CboxConnection(
 {
 }
 
-namespace cbox {
-void connectionStarted(DataOut& out)
-{
-    char header[] = "<!BREWBLOX,";
-    const std::string& version = versionCsv();
-    out.writeBuffer(header, strlen(header));
-    out.writeBuffer(version.data(), version.size());
-    out.write(',');
-    cbox::EncodedDataOut hexOut(out);
-
-#if PLATFORM_ID == 3
-    int resetReason = 0;
-#else
-    auto resetReason = uint8_t(0);
-#endif
-    hexOut.write(resetReason);
-    out.write(',');
-#if PLATFORM_ID == 3
-    int resetData = 0;
-#else
-    auto resetData = uint8_t(0);
-#endif
-    hexOut.write(resetData);
-    out.write(',');
-
-    uint8_t deviceId[12] = {0};
-    hexOut.writeBuffer(deviceId, 12);
-    out.write('>');
-    out.write('\n');
-}
-}
-
 void CboxConnection::start()
 {
     cbox::StreamBufDataOut out_cbox(buffer_out);
