+++ brewblox/blox/compiled_proto/src/proto_version.h
@@ -1,3 +1,3 @@
 #pragma once
-#define COMPILED_PROTO_VERSION "bcaa0787"
-#define COMPILED_PROTO_DATE "2021-08-13"
+#define COMPILED_PROTO_VERSION "3726c7c8"
+#define COMPILED_PROTO_DATE "2021-08-16"
